# contracts/api/ingestion-service.openapi.yaml
components:
  schemas:
    BatchItem:
      properties:
        correlationId:
          type: string
        detail:
          type: string
        eventId:
          type: string
        index:
          format: int32
          type: integer
        memberId:
          type: string
        rejectCode:
          enum:
          - SOURCE_DISABLED
          - UNKNOWN_TYPE
          - TYPE_NOT_ALLOWED
          - INVALID_DATA
          - INVALID_TIME
          - MEMBER_NOT_ACTIVE
          type: string
        status:
          enum:
          - ACCEPTED
          - DUPLICATE
          - REJECTED
          - UNMATCHED
          - INVALID
          type: string
      type: object
    BatchResult:
      properties:
        counts:
          $ref: "#/components/schemas/OutcomeCounts"
        items:
          items:
            $ref: "#/components/schemas/BatchItem"
          type: array
        total:
          format: int32
          type: integer
      type: object
...
    InboundEventRequest:
      properties:
        data:
          $ref: "#/components/schemas/JsonNode"
        id:
          type: string
        source:
          type: string
        specversion:
          type: string
        subject:
          type: string
        time:
          type: string
        type:
          type: string
      type: object
...
    IngestResult:
      properties:
        correlationId:
          type: string
        detail:
          type: string
        eventId:
          type: string
        memberId:
          type: string
        rejectCode:
          enum:
          - SOURCE_DISABLED
          - UNKNOWN_TYPE
          - TYPE_NOT_ALLOWED
          - INVALID_DATA
          - INVALID_TIME
          - MEMBER_NOT_ACTIVE
          type: string
        status:
          enum:
          - ACCEPTED
          - DUPLICATE
          - REJECTED
          - UNMATCHED
          type: string
      type: object
...
    JsonNode:
      properties:
        array:
          type: boolean
        bigDecimal:
          type: boolean
        bigInteger:
          type: boolean
        binary:
          type: boolean
        boolean:
          type: boolean
        container:
          type: boolean
        double:
          type: boolean
        embeddedValue:
          type: boolean
        empty:
          type: boolean
        float:
          type: boolean
        floatingPointNumber:
          type: boolean
        int:
          type: boolean
        integralNumber:
          type: boolean
        long:
          type: boolean
        missingNode:
          type: boolean
        nodeType:
          enum:
          - ARRAY
          - BINARY
          - BOOLEAN
          - MISSING
          - "NULL"
          - NUMBER
          - OBJECT
          - POJO
          - STRING
          type: string
        "null":
          type: boolean
        number:
          type: boolean
        object:
          type: boolean
        pojo:
          type: boolean
        short:
          type: boolean
        string:
          type: boolean
        textual:
          deprecated: true
          type: boolean
        valueNode:
          type: boolean
      type: object
...
    OutcomeCounts:
      properties:
        accepted:
          format: int32
          type: integer
        duplicate:
          format: int32
          type: integer
        invalid:
          format: int32
          type: integer
        rejected:
          format: int32
          type: integer
        unmatched:
          format: int32
          type: integer
      type: object
...
    ProblemDetail:
      properties:
        detail:
          type: string
        instance:
          format: uri
          type: string
        properties:
          additionalProperties: {}
          type: object
        status:
          format: int32
          type: integer
        title:
          type: string
        type:
          format: uri
          type: string
      type: object
...
    TransactionRequest:
      properties:
        amount:
          type: number
        channel:
          type: string
        currency:
          type: string
        items:
          $ref: "#/components/schemas/JsonNode"
        kind:
          type: string
        memberRef:
          type: string
        occurredAt:
          type: string
        orderId:
          type: string
        source:
          type: string
      type: object
info:
  description: "API HTTP di ingestion-service (docs/06 §2). Generata da OpenApiExportIT (M8.8, docs/18 §3.6): non modificare a mano."
  title: Loyalty Hub · ingestion-service
  version: v1
openapi: 3.1.0
paths:
...
  /v1/events:
    post:
      operationId: ingest
      parameters:
      - in: header
        name: X-LH-Reprocess
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/cloudevents+json:
            schema:
              $ref: "#/components/schemas/InboundEventRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/InboundEventRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IngestResult"
          description: OK
      summary: Ingest — events
      tags:
      - events
  /v1/events/batch:
    post:
      operationId: ingestBatch
      requestBody:
        content:
          application/cloudevents-batch+json:
            schema:
              items:
                $ref: "#/components/schemas/InboundEventRequest"
              maxItems: 1000
              type: array
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/InboundEventRequest"
              maxItems: 1000
              type: array
        required: true
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BatchResult"
          description: Accepted
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Corpo non JSON o non array; corpo oltre il tetto in streaming
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "FORBIDDEN_ROLE: serve il ruolo SOURCE (o ADMIN); SOURCE_MISMATCH: un elemento dichiara una fonte diversa dal client src-<codice>, l'intero batch è respinto e nulla è salvato"
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "BATCH_BODY_TOO_LARGE: corpo oltre il tetto in byte"
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "BATCH_EMPTY, BATCH_TOO_LARGE, BATCH_OVER_RATE_LIMIT (il batch da solo supera il limite di eventi al minuto: non ritentabile così com'è, va diviso)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "RATE_LIMITED: il batch starebbe nel limite ma la finestra corrente dell'indirizzo è piena (Retry-After)"
      summary: Ingest batch
      tags:
      - events-batch
...
  /v1/transactions:
    post:
      operationId: transaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IngestResult"
          description: OK
      summary: Transaction — transactions
      tags:
      - transactions