# services/ingestion-service/src/main/java/io/loyaltyhub/ingestion/api/EventsBatchController.java
package io.loyaltyhub.ingestion.api;

import io.loyaltyhub.common.web.ActorHolder;
import io.loyaltyhub.common.web.GlobalExceptionHandler;
import io.loyaltyhub.common.web.LhException;
import io.loyaltyhub.common.web.RequiresRole;
import io.loyaltyhub.common.web.Role;
import io.loyaltyhub.ingestion.application.BatchIngestionService;
import io.loyaltyhub.ingestion.application.SourceBinding;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.parameters.RequestBody;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import jakarta.servlet.http.HttpServletRequest;
import org.springframework.http.HttpEntity;
import org.springframework.http.HttpStatus;
import org.springframework.http.ProblemDetail;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;
import tools.jackson.databind.JsonNode;

import java.net.URI;

/**
 * Ingresso batch (F2-ING-01, docs/18 §3.6; docs/servizi/ingestion-service.md §3): {@code POST /v1/events/batch}, array
 * JSON di al più {@value BatchIngestionService#MAX_EVENTS} CloudEvent (modo <em>batched</em> di CloudEvents), {@code 202}
 * con l'esito di ogni elemento nell'ordine ricevuto ({@link BatchIngestionService}). Il limite di frequenza per
 * indirizzo conta un evento per elemento ({@link IngressRateLimitFilter#admitEvents}); il corpo ha un tetto in byte
 * ({@link BatchBodyLimitFilter}).
 */
@RestController
@RequestMapping("/v1")
public class EventsBatchController {

    private final BatchIngestionService batch;
    private final IngressRateLimitFilter rateLimit;

    public EventsBatchController(BatchIngestionService batch, IngressRateLimitFilter rateLimit) {
        this.batch = batch;
        this.rateLimit = rateLimit;
    }

    /** Finestra dell'indirizzo piena per un batch che altrimenti starebbe nel limite: {@code 429} con {@code Retry-After}. */
    static final class RateLimited extends RuntimeException {
        private final long retryAfterSeconds;

        RateLimited(String detail, long retryAfterSeconds) {
            super(detail);
            this.retryAfterSeconds = retryAfterSeconds;
        }
    }

    // Il corpo arriva come HttpEntity (niente @RequestBody di Spring): l'annotazione @RequestBody qui è quella di
    // OpenAPI e descrive gli elementi come InboundEventRequest; il servizio legge l'albero JSON per dare un esito anche
    // agli elementi malformati.
    @PostMapping(
            path = "/events/batch",
            consumes = {"application/json", "application/cloudevents-batch+json"})
    @ResponseStatus(HttpStatus.ACCEPTED)
    @RequestBody(required = true, content = @Content(array = @ArraySchema(
            maxItems = BatchIngestionService.MAX_EVENTS,
            schema = @Schema(implementation = InboundEventRequest.class))))
    @ApiResponse(responseCode = "400", description = "Corpo non JSON o non array; corpo oltre il tetto in streaming",
            content = @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class)))
    @ApiResponse(responseCode = "403", description = "FORBIDDEN_ROLE: serve il ruolo SOURCE (o ADMIN); SOURCE_MISMATCH: un"
            + " elemento dichiara una fonte diversa dal client src-<codice>, l'intero batch è respinto e nulla è salvato",
            content = @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class)))
    @ApiResponse(responseCode = "413", description = "BATCH_BODY_TOO_LARGE: corpo oltre il tetto in byte",
            content = @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class)))
    @ApiResponse(responseCode = "422", description = "BATCH_EMPTY, BATCH_TOO_LARGE, BATCH_OVER_RATE_LIMIT (il batch da solo"
            + " supera il limite di eventi al minuto: non ritentabile così com'è, va diviso)",
            content = @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class)))
    @ApiResponse(responseCode = "429", description = "RATE_LIMITED: il batch starebbe nel limite ma la finestra"
            + " corrente dell'indirizzo è piena (Retry-After)",
            content = @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class)))
    // Q-492: ingresso delle fonti solo per il ruolo SOURCE (utenza di integrazione, client src-<codice>); ADMIN passa
    // per regola dell'interceptor. La fonte dichiarata deve coincidere con il client (SourceBinding).
    @RequiresRole(Role.SOURCE)
    public ResponseEntity<BatchIngestionService.BatchResult> ingestBatch(HttpEntity<JsonNode> request,
                                                                        HttpServletRequest http) {
        JsonNode body = request.getBody();
        int events = BatchIngestionService.requireBatch(body);
        // Q-492: un solo elemento di un'altra fonte respinge l'intero batch, prima del limite di frequenza e di ogni scrittura.
        SourceBinding.requireMatchAll(ActorHolder.get(), body);
        IngressRateLimitFilter.Admission admission = rateLimit.admitEvents(http, events);
        if (admission.tooMany()) {
            throw LhException.validation("BATCH_OVER_RATE_LIMIT", "Un batch di " + events + " eventi supera il limite di "
                    + rateLimit.perMinute() + " eventi al minuto per questo indirizzo e non passerebbe mai: dividilo in"
                    + " batch di al più " + rateLimit.perMinute() + " eventi.");
        }
        if (!admission.admitted()) {
            throw new RateLimited("Limite di " + rateLimit.perMinute()
                    + " eventi al minuto superato per questo indirizzo: riprova tra poco",
                    Math.max(1, (admission.retryAfterMs() + 999) / 1000));
        }
        return ResponseEntity.status(HttpStatus.ACCEPTED).body(batch.ingest(body));
    }

    @ExceptionHandler(RateLimited.class)
    ResponseEntity<ProblemDetail> onRateLimited(RateLimited e, HttpServletRequest http) {
        ProblemDetail pd = ProblemDetail.forStatusAndDetail(HttpStatus.TOO_MANY_REQUESTS, e.getMessage());
        pd.setType(URI.create(GlobalExceptionHandler.PROBLEM_TYPE_PREFIX + "rate-limited"));
        pd.setTitle("Troppe richieste");
        pd.setInstance(URI.create(http.getRequestURI()));
        pd.setProperty("code", "RATE_LIMITED");
        return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS)
                .header("Retry-After", String.valueOf(e.retryAfterSeconds))
                .body(pd);
    }
}